+++ src/main/java/it/its/testEmployeesDB/repository/ProvincesRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.its.testEmployeesDB.dao.ProvincesDao;
 
-public interface ProvincesRepository extends JpaRepository<ProvincesDao, Long> {
+public interface ProvincesRepository extends JpaRepository<ProvincesDao, Integer> {
 
 }
